Split 'publisher' commands into 'authclient' and 'groups'
Following discussion in [1], split the 'publisher' commands into two
separate sets of commands to manage OAuth clients and groups
respectively.

This means that the CLI commands now map better to the actual entities
in the system.

[1] https://hypothes-is.slack.com/archives/public/p1484904802002707
