+++ h/cli/__init__.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 
 import functools
 import logging
-import os
 
 import click
 from pyramid import paster
@@ -15,14 +14,15 @@ from h import __version__
 log = logging.getLogger('h')
 
 SUBCOMMANDS = (
+    'h.cli.commands.authclient.authclient',
     'h.cli.commands.celery.celery',
     'h.cli.commands.devserver.devserver',
+    'h.cli.commands.groups.groups',
     'h.cli.commands.init.init',
     'h.cli.commands.initdb.initdb',
     'h.cli.commands.migrate.migrate',
     'h.cli.commands.move_uri.move_uri',
     'h.cli.commands.normalize_uris.normalize_uris',
-    'h.cli.commands.publisher.publisher',
     'h.cli.commands.search.search',
     'h.cli.commands.shell.shell',
     'h.cli.commands.user.user',
